fix: decouple container cache logic to containerCache. (#7637)
